+++ jest.config.ts
@@ -4,10 +4,10 @@ module.exports = createConfig('jest', {
   // setupFilesAfterEnv is used after the jest environment has been loaded. In general this is what you want.
   // If you want to add config BEFORE jest loads, use setupFiles instead.
   setupFilesAfterEnv: [
-    '<rootDir>/src/setupTest.js',
+    '<rootDir>/src/setupTest.tsx',
   ],
   coveragePathIgnorePatterns: [
-    'src/setupTest.js',
+    'src/setupTest.tsx',
     'src/i18n',
   ],
 });
